+++ examples/apps/csvql/queries/bad-column.sql
@@ -1,0 +1,1 @@
+SELECT name, salery FROM employees