+++ examples/apps/csvql/queries/bad-limit.sql
@@ -1,0 +1,1 @@
+SELECT name FROM employees LIMIT -3